- added autodetection of encoding using ICU library and KEncodingDetection class as fallback (CMake find module needs more work) - changed InputFormat and OutputFormat classes: they no longer deal with reading data or encoding issues, which is now handled by FormatsManager class - added Russian translation contributed by Alexander Antsev - added support for YouTube Captions subtitle format - moved "change encoding action"" in main toolbar to be next to the open subtitle - fixed precision of time adjust methods - fixed translation of subtitles
git-svn-id: https://subcomposer.svn.sourceforge.net/svnroot/subcomposer/trunk@44 2bc23f0e-8be5-457b-a8bb-2dd17a7f7cc8

+++ src/core/subtitleline.cpp
@@ -693,7 +693,7 @@ void SubtitleLine::shiftTimes( long mseconds )
 		);
 }
 
-void SubtitleLine::adjustTimes( long shiftMseconds, double scaleFactor )
+void SubtitleLine::adjustTimes( double shiftMseconds, double scaleFactor )
 {
 	if ( shiftMseconds || scaleFactor != 1.0 )
 	{
